+++ hypervisor/include/arch/x86/page.h
@@ -8,7 +8,6 @@
 #define PAGE_H
 
 #include <spinlock.h>
-#include <board_info.h>
 
 #define PAGE_SHIFT	12U
 #define PAGE_SIZE	(1U << PAGE_SHIFT)
@@ -29,30 +28,6 @@
 #define PT_PAGE_NUM(size)	(((size) + PDE_SIZE - 1UL) >> PDE_SHIFT)
 
 
-/**
- * @brief Page tables level in IA32 paging mode
- */
-enum _page_table_level {
-        /**
-         * @brief The PML4 level in the page tables
-         */
-	IA32E_PML4 = 0,
-        /**
-         * @brief The Page-Directory-Pointer-Table level in the page tables
-         */
-	IA32E_PDPT = 1,
-        /**
-         * @brief The Page-Directory level in the page tables
-         */
-	IA32E_PD = 2,
-        /**
-         * @brief The Page-Table level in the page tables
-         */
-	IA32E_PT = 3,
-};
-
-struct acrn_vm;
-
 struct page {
 	uint8_t contents[PAGE_SIZE];
 } __aligned(PAGE_SIZE);
@@ -65,16 +40,6 @@ struct page_pool {
 	uint64_t last_hint_id;
 
 	struct page *dummy_page;
-};
-
-struct pgtable {
-	uint64_t default_access_right;
-	struct page_pool *pool;
-	bool (*large_page_support)(enum _page_table_level level, uint64_t prot);
-	uint64_t (*pgentry_present)(uint64_t pte);
-	void (*clflush_pagewalk)(const void *p);
-	void (*tweak_exe_right)(uint64_t *entry);
-	void (*recover_exe_right)(uint64_t *entry);
 };
 
 struct page *alloc_page(struct page_pool *pool);
